requirements.txt
django==3.1
psycopg2-binary==2.8.5
djangorestframework==3.11.1
factory-boy==3.0.1